+++ src/controllers/product.controller.ts
@@ -1,10 +1,15 @@
 /* eslint-disable @typescript-eslint/strict-boolean-expressions */
+/* eslint-disable @typescript-eslint/no-misused-promises */
+import { Types } from 'mongoose'
 import {
+  InvalidIdError,
   MissingFieldError,
   MissingParamError,
+  SomethingWrong,
   UpdateNotAllowedError
 } from '../errors/controllers.errors'
 import ControllerHelper from '../helpers/controller.helper'
+import CategoryModel from '../models/category.model'
 import { ProductServiceAbstract } from '../services/product.service'
 import {
   BaseControllerAbstract,
@@ -32,13 +37,21 @@ export class ProductController implements BaseControllerAbstract {
         }
       }
 
+      const isCategoryValid = await CategoryModel.findById(body.category)
+
+      if (!Types.ObjectId.isValid(body.category) || !isCategoryValid) {
+        return ControllerHelper.badRequest(new InvalidIdError('category'))
+      }
+
       const product = await this.productService.create(body)
       // const productController = new ProductController(new ProductService())
       // productController.create()
 
       return ControllerHelper.created(product)
     } catch (error) {
-      return ControllerHelper.serverError()
+      return ControllerHelper.serverError(
+        new SomethingWrong('Invalid product create data')
+      )
     }
   }
 
@@ -51,11 +64,17 @@ export class ProductController implements BaseControllerAbstract {
         return ControllerHelper.badRequest(new MissingParamError('id'))
       }
 
+      if (!Types.ObjectId.isValid(params.id)) {
+        return ControllerHelper.badRequest(new InvalidIdError('product'))
+      }
+
       const product = await this.productService.getOne(params.id)
 
       return ControllerHelper.ok(product)
     } catch (error) {
-      return ControllerHelper.serverError()
+      return ControllerHelper.serverError(
+        new SomethingWrong('Incorrect id param')
+      )
     }
   }
 
@@ -65,7 +84,9 @@ export class ProductController implements BaseControllerAbstract {
 
       return ControllerHelper.ok(products)
     } catch (error) {
-      return ControllerHelper.serverError()
+      return ControllerHelper.serverError(
+        new SomethingWrong('No products data')
+      )
     }
   }
 
@@ -78,13 +99,17 @@ export class ProductController implements BaseControllerAbstract {
         return ControllerHelper.badRequest(new MissingParamError('id'))
       }
 
+      if (!Types.ObjectId.isValid(params.id)) {
+        return ControllerHelper.badRequest(new InvalidIdError('product'))
+      }
+
       // verficar se os campos fornecidos para atualização são permitidos
       const body = httpRequest.body
       // body = {id: '12', name: 'black shoe'}
       // id / name = keys
       // Object.keys(body) = id, name
 
-      const allowedUpdates = ['name', 'price', 'imageUrl', 'category']
+      const allowedUpdates = ['name', 'price', 'imageUrl']
 
       const receiveUpdateNotAllowed = Object.keys(body).some(
         (update) => !allowedUpdates.includes(update)
@@ -98,7 +123,9 @@ export class ProductController implements BaseControllerAbstract {
 
       return ControllerHelper.ok(product)
     } catch (error) {
-      return ControllerHelper.serverError()
+      return ControllerHelper.serverError(
+        new SomethingWrong('Incorrect id param, invalid product update data')
+      )
     }
   }
 
@@ -111,11 +138,17 @@ export class ProductController implements BaseControllerAbstract {
         return ControllerHelper.badRequest(new MissingParamError('id'))
       }
 
+      if (!Types.ObjectId.isValid(params.id)) {
+        return ControllerHelper.badRequest(new InvalidIdError('product'))
+      }
+
       const product = await this.productService.delete(params.id)
 
       return ControllerHelper.ok(product)
     } catch (error) {
-      return ControllerHelper.serverError()
+      return ControllerHelper.serverError(
+        new SomethingWrong('Incorrect id param')
+      )
     }
   }
 }
